fix build error in Qt6

+++ src/Launchy/Windows/LaunchyWidgetWin.cpp
@@ -27,7 +27,7 @@ LaunchyWidgetWin::LaunchyWidgetWin(CommandFlags command)
     commandMessageId = RegisterWindowMessageW(L"LaunchyCommand");
 }
 
-bool LaunchyWidgetWin::nativeEvent(const QByteArray &eventType, void *message, long *result) {
+bool LaunchyWidgetWin::nativeEvent(const QByteArray& eventType, void* message, intptr_t* result) {
     MSG* msg = (MSG*)message;
     switch (msg->message) {
     case WM_SETTINGCHANGE:
